GurkhaWatch/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.7
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-mathfilters==1.0.0
django-tinymce==3.3.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.0.3
Pillow==8.3.2
python-decouple==3.5
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
stripe==2.61.0
urllib3==1.26.7
